automation/src/requirements.txt
pexpect==4.6.0
pyserial==3.4
PyYAML>=5.3.1
